pyproject.toml
[tool.poetry]
name = "lobbys"
version = "0.1.0"
description = ""
authors = ["Atilio <atiliogoulart@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.6"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
